fix(Application.java): 큰 수 계산 범위 추가

+++ src/main/java/calculator/Application.java
@@ -36,7 +36,7 @@ public class Application {
         long Result = Constant.ZERO;
         for (String s : NumberList) {
             if (s.matches(Constant.MATCHER_CHECK)) {
-                Result += Integer.parseInt(s);
+                Result += Long.parseLong(s);
             } else if (s.matches(Constant.NONE_INPUT)) {
                 Result += Constant.ZERO;
             } else {
